feat: exchange identity extractor
