+++ edc-extensions/dataspace-protocol/cx-dataspace-protocol/src/test/java/org/eclipse/tractusx/edc/protocol/cx/identifier/BpnExtractionFunctionTest.java
@@ -1,4 +1,5 @@
 /*
+ * Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  * Copyright (c) 2025 Cofinity-X GmbH
  * Copyright (c) 2026 SAP SE
  *
@@ -20,50 +21,95 @@
 
 package org.eclipse.tractusx.edc.protocol.cx.identifier;
 
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.CredentialSubject;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.Issuer;
 import org.eclipse.edc.iam.verifiablecredentials.spi.model.VerifiableCredential;
+import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.iam.ClaimToken;
 import org.eclipse.edc.spi.monitor.Monitor;
-import org.eclipse.tractusx.edc.protocol.core.identifier.MembershipCredentialIdExtractionFunction;
-import org.eclipse.tractusx.edc.protocol.core.identifier.MembershipCredentialIdExtractionFunctionTest;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.eclipse.tractusx.edc.edr.spi.CoreConstants.CX_CREDENTIAL_NS;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class BpnExtractionFunctionTest extends MembershipCredentialIdExtractionFunctionTest {
+public class BpnExtractionFunctionTest {
 
+    public static final String DID = "did:web:example";
     public static final String BPN = "bpn";
     public static final String ID_PROPERTY = "holderIdentifier";
 
     private final Monitor monitor = mock();
 
-    @Override
-    protected MembershipCredentialIdExtractionFunction extractionFunction() {
+    private BpnExtractionFunction extractionFunction() {
         when(monitor.withPrefix(anyString())).thenReturn(monitor);
         return new BpnExtractionFunction(monitor);
-    }
-    
-    @Override
-    protected String expectedId() {
-        return BPN;
     }
 
     @ParameterizedTest
     @ArgumentsSource(VerifiableCredentialArgumentProvider.class)
     void apply(VerifiableCredential credential) {
         var id = extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", List.of(credential)).build());
-        assertThat(id).isEqualTo(expectedId());
+        assertThat(id).isEqualTo(BPN);
+    }
+
+    @Test
+    void apply_fails_WhenCredentialNotFound() {
+        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", List.of(vc("FooCredential", Map.of("foo", "bar")))).build()))
+                .isInstanceOf(EdcException.class)
+                .hasMessage("Required credential type 'MembershipCredential' not present in ClaimToken, cannot extract property '%s'", ID_PROPERTY);
+    }
+
+    @Test
+    void apply_fails_whenNoVcClaims() {
+        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().build()))
+                .isInstanceOf(EdcException.class)
+                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken did not contain a 'vc' claim");
+    }
+
+    @Test
+    void apply_fails_whenNullVcClaims() {
+        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", null).build()))
+                .isInstanceOf(EdcException.class)
+                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken did not contain a 'vc' claim");
+    }
+
+    @Test
+    void apply_fails_WhenVcClaimIsNotList() {
+        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", "wrong").build()))
+                .isInstanceOf(EdcException.class)
+                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken contains a 'vc' claim, but the type is incorrect. Expected java.util.List, got java.lang.String.");
+    }
+
+    @Test
+    void apply_fails_WhenVcClaimsIsEmptyList() {
+        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", List.of()).build()))
+                .isInstanceOf(EdcException.class)
+                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken contains a 'vc' claim but it did not contain any VerifiableCredentials.");
+    }
+
+    private static VerifiableCredential vc(String type, Map<String, Object> claims) {
+        return VerifiableCredential.Builder.newInstance().type(type)
+                .issuanceDate(Instant.now())
+                .issuer(new Issuer("issuer", Map.of()))
+                .credentialSubject(CredentialSubject.Builder.newInstance()
+                        .id(claims.containsKey("id") ? claims.get("id").toString() : null)
+                        .claims(claims)
+                        .build())
+                .build();
     }
 
     private static class VerifiableCredentialArgumentProvider implements ArgumentsProvider {
