+++ edc-extensions/dataspace-protocol/dataspace-protocol-core/src/test/java/org/eclipse/tractusx/edc/protocol/core/CoreDataspaceProtocolExtensionTest.java
@@ -21,11 +21,11 @@
 package org.eclipse.tractusx.edc.protocol.core;
 
 import org.eclipse.edc.boot.system.injection.ObjectFactory;
+import org.eclipse.edc.iam.decentralizedclaims.core.defaults.DefaultDcpParticipantIdExtractionFunction;
 import org.eclipse.edc.junit.extensions.DependencyInjectionExtension;
 import org.eclipse.edc.protocol.dsp.http.spi.api.DspBaseWebhookAddress;
 import org.eclipse.edc.protocol.spi.DataspaceProfileContextRegistry;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
-import org.eclipse.tractusx.edc.protocol.core.identifier.DidExtractionFunction;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -63,7 +63,7 @@ class CoreDataspaceProtocolExtensionTest {
                 dataspaceProfileContext -> dataspaceProfileContext.name().equals(DATASPACE_PROTOCOL_HTTP_V_2025_1) &&
                         dataspaceProfileContext.protocolVersion().equals(V_2025_1) &&
                         dataspaceProfileContext.webhook().url().equals(webhook + V_2025_1_PATH) &&
-                        dataspaceProfileContext.idExtractionFunction() instanceof DidExtractionFunction
+                        dataspaceProfileContext.idExtractionFunction() instanceof DefaultDcpParticipantIdExtractionFunction
                 ));
     }
 }
