--- edc-extensions/dataspace-protocol/dataspace-protocol-core/src/testFixtures/java/org/eclipse/tractusx/edc/protocol/core/identifier/MembershipCredentialIdExtractionFunctionTest.java
@@ -1,92 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- * Copyright (c) 2025 Cofinity-X GmbH
- * Copyright (c) 2026 SAP SE
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
-
-package org.eclipse.tractusx.edc.protocol.core.identifier;
-
-import org.eclipse.edc.iam.verifiablecredentials.spi.model.CredentialSubject;
-import org.eclipse.edc.iam.verifiablecredentials.spi.model.Issuer;
-import org.eclipse.edc.iam.verifiablecredentials.spi.model.VerifiableCredential;
-import org.eclipse.edc.spi.EdcException;
-import org.eclipse.edc.spi.iam.ClaimToken;
-import org.junit.jupiter.api.Test;
-
-import java.time.Instant;
-import java.util.List;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-public abstract class MembershipCredentialIdExtractionFunctionTest {
-    
-    public static final String DID = "did:web:example";
-    
-    @Test
-    void apply_fails_WhenCredentialNotFound() {
-        var function = extractionFunction();
-        assertThatThrownBy(() -> function.apply(ClaimToken.Builder.newInstance().claim("vc", List.of(vc("FooCredential", Map.of("foo", "bar")))).build()))
-                .isInstanceOf(EdcException.class)
-                .hasMessage("Required credential type 'MembershipCredential' not present in ClaimToken, cannot extract property '%s'", function.identityProperty());
-    }
-    
-    @Test
-    void apply_fails_whenNoVcClaims() {
-        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().build()))
-                .isInstanceOf(EdcException.class)
-                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken did not contain a 'vc' claim");
-    }
-    
-    @Test
-    void apply_fails_whenNullVcClaims() {
-        
-        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", null).build()))
-                .isInstanceOf(EdcException.class)
-                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken did not contain a 'vc' claim");
-    }
-    
-    @Test
-    void apply_fails_WhenVcClaimIsNotList() {
-        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", "wrong").build()))
-                .isInstanceOf(EdcException.class)
-                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken contains a 'vc' claim, but the type is incorrect. Expected java.util.List, got java.lang.String.");
-    }
-    
-    @Test
-    void apply_fails_WhenVcClaimsIsEmptyList() {
-        assertThatThrownBy(() -> extractionFunction().apply(ClaimToken.Builder.newInstance().claim("vc", List.of()).build()))
-                .isInstanceOf(EdcException.class)
-                .hasMessageContaining("Failed to fetch credentials from the claim token: ClaimToken contains a 'vc' claim but it did not contain any VerifiableCredentials.");
-    }
-    
-    public static VerifiableCredential vc(String type, Map<String, Object> claims) {
-        return VerifiableCredential.Builder.newInstance().type(type)
-                .issuanceDate(Instant.now())
-                .issuer(new Issuer("issuer", Map.of()))
-                .credentialSubject(CredentialSubject.Builder.newInstance()
-                        .id(claims.containsKey("id") ? claims.get("id").toString() : null)
-                        .claims(claims)
-                        .build())
-                .build();
-    }
-    
-    protected abstract MembershipCredentialIdExtractionFunction extractionFunction();
-    
-    protected abstract String expectedId();
-}
